Restrict valid profile names and implement associated unit tests. (#53)
Since the profile name is used as basis for the docker container name it
must be restricted to values valid for containers names. While that
definition might change, the conservative restrictions imposed with this
change set should protect us against any surprises in that regard for
now.

In addition, we do not allow underscores since this might interefere
with the ability to change the tool implementation to use docker-compose
in the future.

Fixes #36.