Setting default app list in profiles empty (#166)
The aiidalab-widgets-base is served as a library to provide the common widgets shared by apps. At the moment, it is an app installed default with the aiidalab-launch. We didn't make the stable release of the package yet which makes the default installed app in-compatible with the container. 
I remove it here for not install the AWB as default but only with the Apps.