Call echo_logs only when in DEBUG mode (#167)
The echo_logs function prints out the logs from the container (i.e. the output of aiidalab-launch logs) while the user waits for the container to start. However, logs are printed with logging.DEBUG so there's no point in calling this function unless this level of logging is enabled (via the -vvv cmdline parameter).