Trigger pull explictly if needed (#184)
fixes #179

The pull option was set as the default that will cause to download the latest version when it is available, which is not desired as discussed in #179. In most cases, users want to keep on using the image they pulled in the first place. The new version of image download has better happened when explicitly triggered by setting pull.

This commit first makes --no-pull the default of --pull/no-pull option. The pull logic tweaks a bit so that when the image does not exist, it will anyway trigger the pull. If the image exists, the --pull needs to be set to trigger the pull.